+++ components/map/controls/component.jsx
