--- components/map/controls/zoom/component.js
@@ -1,71 +1,0 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-import classnames from 'classnames';
-
-// import Icon from 'components/icon';
-import './style.scss';
-
-class ZoomControl extends PureComponent {
-  increaseZoom = (e) => {
-    e.stopPropagation();
-    const { viewport, onClick } = this.props;
-    const { zoom, maxZoom } = viewport;
-
-    onClick(zoom === maxZoom ? zoom : zoom + 1);
-  }
-
-  decreaseZoom = (e) => {
-    e.stopPropagation();
-    const { viewport, onClick } = this.props;
-    const { zoom, minZoom } = viewport;
-
-    onClick(zoom === minZoom ? zoom : zoom - 1);
-  }
-
-  render() {
-    const { className, viewport } = this.props;
-    const { zoom, maxZoom, minZoom } = viewport;
-
-    const classNames = classnames({
-      'c-zoom-control': true,
-      [className]: !!className,
-    });
-
-    const zoomInClass = classnames('zoom-control--btn', { '-disabled': zoom >= maxZoom });
-    const zoomOutClass = classnames('zoom-control--btn', { '-disabled': zoom <= minZoom });
-
-    return (
-      <div className={classNames}>
-        <button
-          className={zoomInClass}
-          type="button"
-          disabled={zoom === maxZoom}
-          onClick={this.increaseZoom}
-        >
-          {/* <Icon name="zoomin" className="-small" /> */}
-        </button>
-        <button
-          className={zoomOutClass}
-          type="button"
-          disabled={zoom === minZoom}
-          onClick={this.decreaseZoom}
-        >
-          {/* <Icon name="zoomout" className="-small" /> */}
-        </button>
-      </div>
-    );
-  }
-}
-ZoomControl.propTypes = {
-  viewport: PropTypes.shape({
-    zoom: PropTypes.number.isRequired,
-    maxZoom: PropTypes.number.isRequired,
-    minZoom: PropTypes.number.isRequired,
-  }).isRequired,
-  className: PropTypes.string,
-  onClick: PropTypes.func.isRequired,
-};
-
-ZoomControl.defaultProps = { className: null };
-
-export default ZoomControl;
