+++ components/map/legend/component.jsx
